# make benchmark to more accurate.