--- a/benchmark/ext/build_data.rb
+++ b/benchmark/ext/build_data.rb
@@ -9,7 +9,9 @@ require_relative '../utils/benchmark_ips_extension'
 h = { 'user' => { id: 1234, name: 'k0kubun' }, book_id: 5432 }
 
 Benchmark.ips do |x|
-  x.report("Faml::AB.build")    { Faml::AttributeBuilder.build("'", true, nil, data: h) }
-  x.report("Hamlit.build_data") { Hamlit::AttributeBuilder.build_data(true, "'", h) }
+  quote = "'"
+  faml_options = { data: h }
+  x.report("Faml::AB.build")    { Faml::AttributeBuilder.build(quote, true, nil, faml_options) }
+  x.report("Hamlit.build_data") { Hamlit::AttributeBuilder.build_data(true, quote, h) }
   x.compare!
 end