@k0kubun 2015-12-12
Ah, I didn't recognize it had cost to build Hash object. Thank you.